+++ packages/gestalt-datepicker/src/DatePicker.tsx
@@ -1,9 +1,33 @@
-import { forwardRef, ReactElement, useEffect, useImperativeHandle, useRef } from 'react';
+import {
+  forwardRef,
+  Fragment,
+  ReactElement,
+  useEffect,
+  useImperativeHandle,
+  useRef,
+  useState,
+} from 'react';
 import { Locale } from 'date-fns/locale';
-import { useGlobalEventsHandler } from 'gestalt';
+import {
+  Button,
+  Flex,
+  Layer,
+  SheetMobile,
+  useDefaultLabel,
+  useDeviceType,
+  useGlobalEventsHandler,
+} from 'gestalt';
 import InternalDatePicker from './DatePicker/InternalDatePicker';
 
+interface Indexable {
+  index(): number;
+}
+
 export type Props = {
+  /**
+   * DatePicker can adapt to mobile devices to [SheetMobile](https://gestalt.pinterest.systems/web/sheetmobile). Mobile adaptation is disabled by default. Set to 'false' to enable SheetMobile in mobile devices. See the [mobile variant](https://gestalt.pinterest.systems/web/datepicker#Mobile) to learn more.
+   */
+  disableMobileUI?: boolean;
   /**
    *  When disabled, DatePicker looks inactive and cannot be interacted with. See the [disabled example](https://gestalt.pinterest.systems/web/datepicker#States) to learn more.
    */
@@ -67,6 +91,14 @@ export type Props = {
    */
   placeholder?: string;
   /**
+   * Callback fired when SheetMobile's in & out animations end. See [SheetMobile's animation variant](https://gestalt.pinterest.systems/web/sheetmobile#Animation) to learn more.
+   */
+  mobileOnAnimationEnd?: (arg1: { animationState: 'in' | 'out' }) => void;
+  /**
+   * An object representing the zIndex value of the SheetMobile where DatePicker is built upon on mobile. Learn more about [zIndex classes](https://gestalt.pinterest.systems/web/zindex_classes)
+   */
+  mobileZIndex?: Indexable;
+  /**
    * Required for date range selection. End date on a date range selection. See the [date range example](https://gestalt.pinterest.systems/web/datepicker#Date-range) to learn more.
    */
   rangeEndDate?: Date | null;
@@ -107,6 +139,7 @@ export type Props = {
 const DatePickerWithForwardRef = forwardRef<HTMLInputElement, Props>(function DatePicker(
   {
     disabled,
+    disableMobileUI = false,
     errorMessage,
     excludeDates,
     helperText,
@@ -117,6 +150,8 @@ const DatePickerWithForwardRef = forwardRef<HTMLInputElement, Props>(function Da
     localeData,
     maxDate,
     minDate,
+    mobileZIndex,
+    mobileOnAnimationEnd,
     name,
     nextRef,
     onChange,
@@ -139,9 +174,114 @@ const DatePickerWithForwardRef = forwardRef<HTMLInputElement, Props>(function Da
     datePickerHandlers: undefined,
   };
 
+  const { accessibilityDismissButtonLabel, dismissButton } = useDefaultLabel('DatePicker');
+
+  const [showMobileCalendar, setShowMobileCalendar] = useState<boolean>(false);
+
+  const deviceType = useDeviceType();
+  const isMobile = deviceType === 'mobile';
+
   useEffect(() => {
     if (datePickerHandlers?.onRender) datePickerHandlers?.onRender();
   }, [datePickerHandlers]);
+
+  if (isMobile && !disableMobileUI) {
+    return (
+      <Fragment>
+        <InternalDatePicker
+          ref={innerInputRef}
+          disabled={disabled}
+          errorMessage={errorMessage}
+          excludeDates={excludeDates}
+          helperText={helperText}
+          id={id}
+          idealDirection={idealDirection}
+          includeDates={includeDates}
+          inputOnly
+          label={label}
+          localeData={localeData}
+          maxDate={maxDate}
+          minDate={minDate}
+          name={name}
+          nextRef={nextRef}
+          onChange={onChange}
+          onFocus={() => setShowMobileCalendar(true)}
+          placeholder={placeholder}
+          rangeEndDate={rangeEndDate}
+          rangeSelector={rangeSelector}
+          rangeStartDate={rangeStartDate}
+          readOnly={readOnly}
+          selectLists={selectLists}
+          value={value}
+        />
+        {showMobileCalendar ? (
+          <Layer zIndex={mobileZIndex}>
+            <SheetMobile
+              footer={
+                <SheetMobile.DismissingElement>
+                  {({ onDismissStart }) => (
+                    <Flex
+                      alignItems="center"
+                      direction="column"
+                      gap={4}
+                      justifyContent="center"
+                      width="100%"
+                    >
+                      <Button
+                        accessibilityLabel={accessibilityDismissButtonLabel}
+                        color="gray"
+                        onClick={() => onDismissStart()}
+                        size="lg"
+                        text={dismissButton}
+                      />
+                    </Flex>
+                  )}
+                </SheetMobile.DismissingElement>
+              }
+              heading=""
+              onAnimationEnd={mobileOnAnimationEnd}
+              onDismiss={() => setShowMobileCalendar(false)}
+              padding="none"
+              showDismissButton={false}
+              size="auto"
+            >
+              <SheetMobile.DismissingElement>
+                {({ onDismissStart }) => (
+                  <Flex
+                    alignItems="center"
+                    direction="column"
+                    gap={4}
+                    justifyContent="center"
+                    width="100%"
+                  >
+                    <InternalDatePicker
+                      errorMessage={errorMessage}
+                      excludeDates={excludeDates}
+                      id={id}
+                      idealDirection={idealDirection}
+                      includeDates={includeDates}
+                      inline
+                      localeData={localeData}
+                      maxDate={maxDate}
+                      minDate={minDate}
+                      nextRef={nextRef}
+                      onChange={onChange}
+                      onSelect={() => onDismissStart()}
+                      rangeEndDate={rangeEndDate}
+                      rangeSelector={rangeSelector}
+                      rangeStartDate={rangeStartDate}
+                      selectLists={selectLists}
+                      value={value}
+                    />
+                  </Flex>
+                )}
+              </SheetMobile.DismissingElement>
+            </SheetMobile>
+          </Layer>
+        ) : null}
+      </Fragment>
+    );
+  }
 
   return (
     <InternalDatePicker